requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.4
cleanco==2.3
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
filelock==3.25.0
fsspec==2026.2.0
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-pptx==1.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.0
shellingham==1.5.4
smart_open==7.5.1
spacy==3.8.11
spacy-alignments==0.9.2
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
srsly==2.5.2
sympy==1.14.0
thinc==8.3.10
tokenizers==0.21.4
torch==2.10.0
tqdm==4.67.3
transformers==4.49.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
xlsxwriter==3.2.9
